+++ apps/wallet/instance/src/main/java/com/tonapps/tonkeeper/fragment/trade/exchange/vm/ExchangeEvent.kt
@@ -1,6 +1,7 @@
 package com.tonapps.tonkeeper.fragment.trade.exchange.vm
 
 import com.tonapps.tonkeeper.fragment.trade.domain.model.ExchangeDirection
+import java.math.BigDecimal
 
 sealed class ExchangeEvent {
     data class NavigateToPickOperator(
@@ -8,7 +9,7 @@ sealed class ExchangeEvent {
         val paymentMethodName: String,
         val country: String,
         val currencyCode: String,
-        val amount: Float,
+        val amount: BigDecimal,
         val direction: ExchangeDirection
     ) : ExchangeEvent()
 }
